+++ packages/ui/src/components/general-animations/index.js
@@ -1,0 +1,2 @@
+export { default as SlideUp } from './slide-up/slide-up';
+export { default as TintOverlay } from './tint-overlay/tint-overlay';