+++ packages/ui/src/components/text-animations/letter-spacing-reveal/letter-spacing-reveal.jsx
@@ -1,107 +1,98 @@
 import PropTypes from 'prop-types';
 import { motion } from 'framer-motion';
-import VisibilitySensor from 'react-visibility-sensor';
-import { useVisibilitySensorWithResetDelay } from '@hooks/interactions';
 import localStyles from './letter-spacing-reveal.module.scss';
 
 /**
  * A text reveal component which will fade in while increasing or decreasing the letter spacing.
  *
- * The text reveal will only occur once the component is visible in the viewport.
+ * This component utilizes the following variant names: `visible` (for when the component should be visible on the
+ * screen) and `hidden` (for when the component should be hidden off the screen). These variants can be used
+ * in parent components to support orchestration such as staggered animations. If you are going to be using these
+ * variants in parent components, you can simply leave the `activated` prop as `undefined` and the component will
+ * automatically sync with the parent component's variants (as long as they are named the same).
  */
-const LetterSpacingReveal = ({
+function LetterSpacingReveal({
     children,
     animationSpeed,
-    resetDelay,
+    activated,
     startLetterSpacing,
     endLetterSpacing,
     // Pass-thru props
     ...passThruProps
-}) => {
-
-    //#region Constants
+}) {
+    // #region Constants
 
     /**
      * The variants for the animation.
      */
     const animationVariants = {
         hidden: {
-            opacity: 0,
+            opacity:       0,
             letterSpacing: `${startLetterSpacing}px`,
-            transition: {
-                duration: 0,
+            transition:    {
+                duration: animationSpeed,
             },
         },
         visible: {
-            opacity: 1,
+            opacity:       1,
             letterSpacing: `${endLetterSpacing}px`,
-            transition: {
+            transition:    {
                 duration: animationSpeed,
             },
         },
     };
 
-    //#endregion
+    // #endregion
 
-    //#region State
+    // #region Effects
 
-    /**
-     * Track the visibility of the component with a reset delay.
-     */
-    const [isVisible, onVisibilityChanged] = useVisibilitySensorWithResetDelay(resetDelay);
+    // #endregion
 
-    //#endregion
+    // #region Functions
 
-    //#region Effects
+    // #endregion
 
-    //#endregion
-
-    //#region Functions
-
-    //#endregion
-
-    //#region Render Functions
+    // #region Render Functions
 
     /**
      * Main render.
      */
     return (
-        <VisibilitySensor onChange={onVisibilityChanged}>
-            <motion.div
-                {...passThruProps}
-                className={`
-                    ${localStyles.letterSpacingReveal}
-                    ${passThruProps?.className}
-                `}
-                variants={animationVariants}
-                initial="hidden"
-                animate={isVisible ? 'visible' : 'hidden'}
-            >
-                {children}
-            </motion.div>
-        </VisibilitySensor>
+        <motion.div
+            {...passThruProps}
+            className={`
+                ${localStyles.letterSpacingReveal}
+                ${passThruProps?.className}
+            `}
+            variants={animationVariants}
+            initial={activated === undefined ? undefined : 'hidden'}
+            animate={
+                activated === undefined
+                    ? undefined
+                    : activated ? 'visible' : 'hidden'
+            }
+        >
+            {children}
+        </motion.div>
     );
 
-    //#endregion
-};
+    // #endregion
+}
 
 LetterSpacingReveal.propTypes = {
     /**
      * The children to render as part of the letter spacing reveal. Only text which doesn't have a letter spacing override
      * within children will be affected by the letter spacing reveal.
      */
-    children: PropTypes.any.isRequired,
+    children:           PropTypes.any.isRequired,
     /**
      * The number of seconds it should take the animation to complete.
      */
-    animationSpeed: PropTypes.number,
+    animationSpeed:     PropTypes.number,
     /**
-     * The number of seconds to delay the reset of the text reveal when it is no longer visible. If the text
-     * is visible again before the reset delay has elapsed, then the reset will be canceled. If this is set to
-     * `0`, then the reset will happen immediately. If this is set to `Infinity`, then the reset will never
-     * happen.
+     * Flag indicating whether the animation should be activated.
      */
-    resetDelay: PropTypes.number,
+    activated:          PropTypes.bool,
     /**
      * The starting letter spacing, in pixels, to use for the animation.
      */
@@ -109,21 +100,21 @@ LetterSpacingReveal.propTypes = {
     /**
      * The ending letter spacing, in pixels, to use for the animation.
      */
-    endLetterSpacing: PropTypes.number,
+    endLetterSpacing:   PropTypes.number,
     /**
      * Any additional props to pass through to the internal div used to wrap the children.
      *
      * This is not a prop of `passThruProps` -- this is simply a representation of any
      * additional props passed to the `LetterSpacingReveal` component that aren't covered above.
      */
-    "...passThruProps": PropTypes.any,
+    '...passThruProps': PropTypes.any,
 };
 
 LetterSpacingReveal.defaultProps = {
-    animationSpeed: 0.75,
-    resetDelay: Infinity,
+    animationSpeed:     0.75,
+    activated:          undefined,
     startLetterSpacing: -4,
-    endLetterSpacing: 1,
+    endLetterSpacing:   1,
 };
 
 export default LetterSpacingReveal;