+++ packages/ui/src/components/text-animations/letter-spacing-reveal/letter-spacing-reveal.stories.jsx
@@ -2,6 +2,7 @@ import {
     setAsCategory,
     setAsDisabled,
 } from '@unifire-js/storybook-utils';
+import { motion } from 'framer-motion';
 import LetterSpacingReveal from './letter-spacing-reveal';
 
 // Construct the argTypes object
@@ -9,9 +10,11 @@ const argTypes = {};
 setAsCategory(argTypes, 'UI', [
     'children',
     'animationSpeed',
-    'resetDelay',
     'startLetterSpacing',
     'endLetterSpacing',
+]);
+setAsCategory(argTypes, 'Controls', [
+    'activated',
 ]);
 setAsCategory(argTypes, 'Pass Thru Props', [
     '...passThruProps',
@@ -22,48 +25,74 @@ setAsDisabled(argTypes, [
 
 // Storybook default export
 export default {
-    title: 'text animations/LetterSpacingReveal',
+    title:     'text animations/LetterSpacingReveal',
     component: LetterSpacingReveal,
     argTypes,
 };
 
-const Template = (args) => {
+function Template(args) {
     return (
-        <div style={{ width: 'fit-content', height: 'fit-content', padding: '20px', margin: '40px', borderRadius: '10px', background: 'white' }}>
-            <LetterSpacingReveal {...args}/>
-            <div style={{ marginBottom: '1200px' }}>Scroll down to see more.</div>
-            <LetterSpacingReveal {...args}/>
+        <div
+            style={{
+                width: '500px', height: 'fit-content', padding: '20px', margin: '40px', borderRadius: '10px', background: 'white',
+            }}
+        >
+            <LetterSpacingReveal {...args} />
         </div>
     );
-};
+}
+
+function GroupTemplate(args) {
+    const variants = {
+        visible: {
+            transition: {
+                when:            'beforeChildren',
+                staggerChildren: 0.3,
+            },
+        },
+        hidden: {
+            transition: {
+                when:             'afterChildren',
+                staggerChildren:  0.3,
+                staggerDirection: -1,
+            },
+        },
+    };
+
+    return (
+        <motion.div
+            style={{
+                width: '500px', height: 'fit-content', padding: '20px', margin: '40px', borderRadius: '10px', background: 'white',
+            }}
+            variants={variants}
+            initial="hidden"
+            animate={args.activated ? 'visible' : 'hidden'}
+        >
+            <LetterSpacingReveal {...args} activated={undefined} />
+            <LetterSpacingReveal {...args} activated={undefined} />
+            <LetterSpacingReveal {...args} activated={undefined} />
+            <LetterSpacingReveal {...args} activated={undefined} />
+            <LetterSpacingReveal {...args} activated={undefined} />
+            <LetterSpacingReveal {...args} activated={undefined} />
+        </motion.div>
+    );
+}
 
 // Basic Demo
 export const Basic = Template.bind({});
 Basic.args = {
-    children: 'Hello World',
+    children:  'Hello World',
+    activated: true,
 };
 
 // Style Override Demo
 export const StyleOverride = Template.bind({});
 StyleOverride.args = {
     children: 'Hello World',
-    style: {
+    style:    {
         fontSize: '24px',
     },
-};
-
-// Reset Demo
-export const Reset = Template.bind({});
-Reset.args = {
-    children: 'Hello World',
-    resetDelay: 2,
-};
-
-// No Reset Delay Demo
-export const NoResetDelay = Template.bind({});
-NoResetDelay.args = {
-    children: 'Hello World',
-    resetDelay: 0,
+    activated: true,
 };
 
 // Complex Children Demo
@@ -78,12 +107,21 @@ ComplexChildren.args = {
             <p>How are you today?</p>
         </>
     ),
+    activated: true,
 };
 
 // Invert Letter Spacing Animation Demo
 export const InvertLetterSpacingAnimation = Template.bind({});
 InvertLetterSpacingAnimation.args = {
-    children: 'Hello World',
-    endLetterSpacing: 1,
+    children:           'Hello World',
+    endLetterSpacing:   1,
     startLetterSpacing: 8,
+    activated:          true,
+};
+
+// Group Demo
+export const Group = GroupTemplate.bind({});
+Group.args = {
+    children:  'Hello World',
+    activated: true,
 };