+++ packages/ui/src/components/text-animations/typing-reveal/typing-reveal.stories.jsx
@@ -10,7 +10,9 @@ setAsCategory(argTypes, 'UI', [
     'lines',
     'typingSpeed',
     'fadeSpeed',
-    'resetDelay',
+]);
+setAsCategory(argTypes, 'Controls', [
+    'activated',
 ]);
 setAsCategory(argTypes, 'Pass Thru Props', [
     '...passThruProps',
@@ -21,20 +23,22 @@ setAsDisabled(argTypes, [
 
 // Storybook default export
 export default {
-    title: 'text animations/TypingReveal',
+    title:     'text animations/TypingReveal',
     component: TypingReveal,
     argTypes,
 };
 
-const Template = (args) => {
+function Template(args) {
     return (
-        <div style={{ width: 'fit-content', height: 'fit-content', padding: '20px', margin: '40px', borderRadius: '10px', background: 'white' }}>
-            <TypingReveal {...args}/>
-            <div style={{ marginBottom: '1200px' }}>Scroll down to see more.</div>
-            <TypingReveal {...args}/>
+        <div
+            style={{
+                width: 'fit-content', height: 'fit-content', padding: '20px', margin: '40px', borderRadius: '10px', background: 'white',
+            }}
+        >
+            <TypingReveal {...args} />
         </div>
     );
-};
+}
 
 // Basic Demo
 export const Basic = Template.bind({});
@@ -43,6 +47,7 @@ Basic.args = {
         "'Why waste time say lot word",
         "when few word do trick...'",
     ],
+    activated: true,
 };
 
 // Style Override Demo
@@ -53,27 +58,8 @@ StyleOverride.args = {
         "when few word do trick...'",
     ],
     style: {
-        fontSize: '24px',
+        fontSize:   '24px',
         alignItems: 'center',
     },
-};
-
-// Reset Demo
-export const Reset = Template.bind({});
-Reset.args = {
-    lines: [
-        "'Why waste time say lot word",
-        "when few word do trick...'",
-    ],
-    resetDelay: 2,
-};
-
-// No Reset Delay Demo
-export const NoResetDelay = Template.bind({});
-NoResetDelay.args = {
-    lines: [
-        "'Why waste time say lot word",
-        "when few word do trick...'",
-    ],
-    resetDelay: 0,
+    activated: true,
 };